+++ src/main/java/com/datastax/kawoosh/analyser/rules/NbOfSSTablesCheckRule.java
@@ -4,6 +4,6 @@ import com.datastax.kawoosh.analyser.ClusterConfigRetriver;
 
 public class NbOfSSTablesCheckRule extends TableValueInRangeRule {
     public NbOfSSTablesCheckRule(ClusterConfigRetriver clusterConfigRetriver) {
-        super(clusterConfigRetriver, "SSTable Count", "SSTable count", (Long) 0L,(Long) 40L);
+        super(clusterConfigRetriver, "SSTable Count", "SSTable count",  0d, 40d);
     }
 }
